requirements.txt
python-docx
webcolors
textwrap3
Pillow